Add tests_brainstorm__analysis.py
This adds and rests two prototype functions `retrieve(...)` and
`interview_persona(...)`.

`retrieve(...)` is a version of memory-retrieval function
`new_retrieve(...)` defined in `retrieve.py`. The latter function
contains some experiments with different weighting in retrieval-score
calculation from (recency,importance,relevance) scores. The version in
`retrieve(...)` version specifies the weights used in the
generative-agents paper.

`interview_persona(...)` is a version of `open_convo_session(...)`
defined in `converse.py`. The latter function loops over the results of
`input(...)` to allow the user to converse with the persona until the
user types "end_convo". The version in `interview_persona(...)` performs
a single conversational exchange and is for use in testing and analysis.