Add test_brainstorm__prototype__persona_ablations__observation_planning_reflection
This brainstorm prototypes a procedure to interview a persona under full
ablation: no observation, planning, or reflection memories.

To make this work I had to insert a single minimal memory of the form
`'{persona.name} is {persona.name}'`. The reason for doing this is that
without at least one event or thought memory an exception is raised:
`ValueError: min() arg is an empty sequence'.

An explanation of the error is written into the comments above
`test_brainstorm__persona_initial_memories(...)` where I examine the
first event, thought, and chat memories in order to figure out how to
setup the minimal memory.