Adjust LangChain cache setup in reverie.py.
This work was part of an effort to try to not hardwire LangChain into
Reverie. On reflection, while I think this is a good idea, I should be
working on other things right now. So I'll try to save the work, set it
aside, and come back to it later.