Add `Conditions.interview__roleplay(...)`.
This method calls LLM to roleplay how a persona would repond to the
given question. This is in order to provide the baseline response for
the ablation conditions per the generative agents paper.

Note that in the paper, a human performs this roleplay. Because I don't
have the resources for this I'm using GPT-3.5-turbo-16k as a proxy. I
want to move to GPT-4, but currently the persona's memory stream is too
large to fit into the GPT-4 context limit.